+++ builder/builder.go
@@ -1,17 +1,16 @@
 package builder
 
 import (
-	"github.com/docker/docker/builder/evaluator"
 	"github.com/docker/docker/runconfig"
 )
 
-// Create a new builder.
-func NewBuilder(opts *evaluator.BuildOpts) *evaluator.BuildFile {
-	return &evaluator.BuildFile{
+// Create a new builder. See
+func NewBuilder(opts *BuildOpts) *BuildFile {
+	return &BuildFile{
 		Dockerfile:    nil,
 		Config:        &runconfig.Config{},
 		Options:       opts,
-		TmpContainers: evaluator.UniqueMap{},
-		TmpImages:     evaluator.UniqueMap{},
+		TmpContainers: UniqueMap{},
+		TmpImages:     UniqueMap{},
 	}
 }
