+++ builder/dispatchers.go
@@ -1,4 +1,4 @@
-package evaluator
+package builder
 
 // This file contains the dispatchers for each command. Note that
 // `nullDispatch` is not actually a command, but support for commands we parse
